feat(drawsystem): add result package and move Result classes
add result package into drawsystem an move DrawReult and
SingleDrawResult into it
